1.7.1 - Bugfixes, performance, Lights Off support (#67)
* Use size_t instead of int more when possible

* Fix #63

* Avoid passing args unnecessarily

* Fix #41 improve split performance

* Improve QQQQ expander code

* Work starting on cleaning up components

* Starting work on making cleaner jacks

* More widget cleanups

* Better code for lit jacks

* New jack code, still need implemented in modules

* First module modernized

* All ported to the new widget library

* Improve naming consistency

* More technical debt repayment

* More support for Lights Off in Darius & knobs

* Make components graphics WTFPL for simplicity

* More work on components & Lights Off

* Buttons are lit now

* Checkpoint, about to make a bulk change...

* Another checkpoint

* Nasty crash on undo fixed!

* Looks like we got full support of Lights Off now

* Quale tonic UI - no underlying support yet

* Custom data structure for PES

* Quale Channel 1 as root note implemented

* Darius: no longer audio rate

* QQQQ: No longer forwards the tonic for now.

* Fix #53 QQQQ Recall

* Clean up for release

+++ src/prng.hpp
@@ -48,11 +48,11 @@ struct prng {
     void init(float seed1, float seed2){
         s[0] = seed1 * 572376460694501; // Keyboard smash - salting seems to improve results
         s[1] = seed2 * 645624357248923;
-        for (int i = 0; i < 50; i++) next(); // Warm up for better results
+        for (size_t i = 0; i < 50; i++) next(); // Warm up for better results
     }
 
     float uniform() {
-        for (int i = 0; i < 50; i++) next(); // More dry runs.
+        for (size_t i = 0; i < 50; i++) next(); // More dry runs.
         return (next() >> (64 - 24)) / std::pow(2.f, 24);
     }
